sample: bump to Debian 10

+++ samples/case_1/run.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 test -d venv || virtualenv -p python3 venv
 source venv/bin/activate
-pip install virt-lightning ansible
+test -f venv/bin/ansible || pip install virt-lightning ansible
 ansible-playbook playbook.yml -vvvv -i inventory.yaml
